+++ src/utils/index.ts
@@ -9,7 +9,7 @@ export * from './scrollbar-width';
 export * from './selection';
 export * from './translate';
 export * from './visibility-observer';
-export * from './debounce';
+export * from './throttle';
 export * from './sort';
 export * from './row-height-cache';
 export * from './column-helper';