+++ docs/changelog.md
@@ -6,6 +6,7 @@
 - Style: Remove row active/select state when in cell mode of material theme
 - Bug: Force fill column mode w/ scrollbarH false allowing bleed when resized
 - Perf: Faster render time by adding css class via host
+- Perf: Debounce window resize calcs
 
 ## 2.0.0
 - BREAKING: Rename from `datatable` to `swui-datatable` (#320)